backend/requirements.txt
fastapi
uvicorn[standard]
openai
anthropic
google-generativeai
python-dotenv
pydantic
jsonschema
